Move scorers and negotiators to proposal module; Improve imports
