+++ examples/managers/proposal_plugins.py
@@ -1,19 +1,26 @@
 import asyncio
 import logging.config
 from datetime import timedelta
-from random import randint
+from random import randint, random
 
 from golem.managers import (
     ActivityPoolManager,
     AddChosenPaymentPlatform,
-    BlacklistProviderId,
+    BlacklistProviderIdNegotiator,
+    BlacklistProviderIdPlugin,
+    Buffer,
     DefaultAgreementManager,
+    DefaultProposalManager,
     LinearAverageCostPricing,
+    MapScore,
+    NegotiatingPlugin,
     PayAllPaymentManager,
+    PropertyValueLerpScore,
+    RandomScore,
     RefreshingDemandManager,
     RejectIfCostsExceeds,
     RejectProposal,
-    SequentialNegotiationManager,
+    ScoringBuffer,
     SequentialWorkManager,
     WorkContext,
     WorkResult,
@@ -22,7 +29,7 @@ from golem.managers import (
     work_plugin,
 )
 from golem.node import GolemNode
-from golem.payload import RepositoryVmPayload
+from golem.payload import RepositoryVmPayload, defaults
 from golem.resources import DemandData, ProposalData
 from golem.utils.logging import DEFAULT_LOGGING
 
@@ -67,34 +74,52 @@ async def main():
         payment_manager.get_allocation,
         payload,
     )
-    negotiation_manager = SequentialNegotiationManager(
+    proposal_manager = DefaultProposalManager(
         golem,
         demand_manager.get_initial_proposal,
         plugins=[
-            AddChosenPaymentPlatform(),
-            # class based plugin
-            BlacklistProviderId(BLACKLISTED_PROVIDERS),
-            RejectIfCostsExceeds(1, linear_average_cost),
-            # func plugin
-            blacklist_func,
-            # lambda plugin
-            lambda _, proposal_data: proposal_data.issuer_id not in BLACKLISTED_PROVIDERS,
-            # lambda plugin with reject reason
-            lambda _, proposal_data: RejectProposal(f"Blacklisting {proposal_data.issuer_id}")
-            if proposal_data.issuer_id in BLACKLISTED_PROVIDERS
-            else None,
+            Buffer(
+                min_size=10,
+                max_size=1000,
+                concurrency_size=5,
+            ),
+            BlacklistProviderIdPlugin(BLACKLISTED_PROVIDERS),
+            NegotiatingPlugin(
+                negotiators=[
+                    AddChosenPaymentPlatform(),
+                    # class based plugin
+                    BlacklistProviderIdNegotiator(BLACKLISTED_PROVIDERS),
+                    RejectIfCostsExceeds(1, linear_average_cost),
+                    # func plugin
+                    blacklist_func,  # type: ignore[list-item]
+                    # lambda plugin
+                    lambda _, proposal_data: proposal_data.issuer_id  # type: ignore[list-item]
+                    not in BLACKLISTED_PROVIDERS,
+                    # lambda plugin with reject reason
+                    lambda _, proposal_data: RejectProposal(  # type: ignore[list-item]
+                        f"Blacklisting {proposal_data.issuer_id}"
+                    )
+                    if proposal_data.issuer_id in BLACKLISTED_PROVIDERS
+                    else None,
+                ]
+            ),
+            ScoringBuffer(
+                min_size=3,
+                max_size=5,
+                concurrency_size=3,
+                scorers=[
+                    MapScore(linear_average_cost, normalize=True, normalize_flip=True),
+                    [0.5, PropertyValueLerpScore(defaults.INF_MEM, zero_at=1, one_at=8)],
+                    [0.1, RandomScore()],
+                    [0.0, lambda proposals_data: [random() for _ in range(len(proposals_data))]],
+                    [0.0, MapScore(lambda proposal_data: random())],
+                ],
+            ),
         ],
     )
     agreement_manager = DefaultAgreementManager(
         golem,
-        negotiation_manager.get_draft_proposal,
-        # plugins=[
-        #     MapScore(linear_average_cost, normalize=True, normalize_flip=True),
-        #     [0.5, PropertyValueLerpScore(defaults.INF_MEM, zero_at=1, one_at=8)],
-        #     [0.1, RandomScore()],
-        #     [0.0, lambda proposals_data: [random() for _ in range(len(proposals_data))]],
-        #     [0.0, MapScore(lambda proposal_data: random())],
-        # ],
+        proposal_manager.get_draft_proposal,
     )
     activity_manager = ActivityPoolManager(golem, agreement_manager.get_agreement, size=3)
     work_manager = SequentialWorkManager(
@@ -106,7 +131,7 @@ async def main():
     )
 
     async with golem:
-        async with payment_manager, demand_manager, negotiation_manager, agreement_manager, activity_manager:  # noqa: E501 line too long
+        async with payment_manager, demand_manager, proposal_manager, agreement_manager, activity_manager:  # noqa: E501 line too long
             await asyncio.sleep(10)
             result: WorkResult = await work_manager.do_work(commands_work_example)
             print(f"\nWORK MANAGER RESULT:{result}\n")
