+++ examples/managers/ssh.py
@@ -5,10 +5,12 @@ import string
 from uuid import uuid4
 
 from golem.managers import (
+    AddChosenPaymentPlatform,
     DefaultAgreementManager,
+    DefaultProposalManager,
+    NegotiatingPlugin,
     PayAllPaymentManager,
     RefreshingDemandManager,
-    SequentialNegotiationManager,
     SequentialWorkManager,
     SingleNetworkManager,
     SingleUseActivityManager,
@@ -84,10 +86,16 @@ async def main():
         payment_manager.get_allocation,
         payload,
     )
-    negotiation_manager = SequentialNegotiationManager(golem, demand_manager.get_initial_proposal)
+    proposal_manager = DefaultProposalManager(
+        golem,
+        demand_manager.get_initial_proposal,
+        plugins=[
+            NegotiatingPlugin(negotiators=[AddChosenPaymentPlatform()]),
+        ],
+    )
     agreement_manager = DefaultAgreementManager(
         golem,
-        negotiation_manager.get_draft_proposal,
+        proposal_manager.get_draft_proposal,
     )
 
     ssh_handler = SshHandler(golem._api_config.app_key, network_manager=network_manager)
@@ -98,7 +106,7 @@ async def main():
         on_activity_start=ssh_handler.on_activity_start,
     )
     work_manager = SequentialWorkManager(golem, activity_manager.do_work)
-    async with golem, network_manager, payment_manager, demand_manager, negotiation_manager, agreement_manager:  # noqa: E501 line too long
+    async with golem, network_manager, payment_manager, demand_manager, proposal_manager, agreement_manager:  # noqa: E501 line too long
         task = asyncio.create_task(work_manager.do_work(ssh_handler.work))
         while not ssh_handler.started:
             await asyncio.sleep(0.1)
