+++ golem/managers/agreement/__init__.py
@@ -1,13 +1,7 @@
 from golem.managers.agreement.default import DefaultAgreementManager
 from golem.managers.agreement.events import AgreementReleased
-from golem.managers.agreement.plugins import MapScore, PropertyValueLerpScore, RandomScore
-from golem.managers.agreement.pricings import LinearAverageCostPricing
 
 __all__ = (
     "AgreementReleased",
     "DefaultAgreementManager",
-    "MapScore",
-    "PropertyValueLerpScore",
-    "RandomScore",
-    "LinearAverageCostPricing",
 )
