--- golem/managers/negotiation/__init__.py
@@ -1,13 +1,0 @@
-from golem.managers.negotiation.plugins import (
-    AddChosenPaymentPlatform,
-    BlacklistProviderId,
-    RejectIfCostsExceeds,
-)
-from golem.managers.negotiation.sequential import SequentialNegotiationManager
-
-__all__ = (
-    "SequentialNegotiationManager",
-    "AddChosenPaymentPlatform",
-    "BlacklistProviderId",
-    "RejectIfCostsExceeds",
-)
