+++ golem/managers/proposal/__init__.py
@@ -1,3 +1,31 @@
 from golem.managers.proposal.default import DefaultProposalManager
+from golem.managers.proposal.plugins import (
+    AddChosenPaymentPlatform,
+    BlacklistProviderIdNegotiator,
+    BlacklistProviderIdPlugin,
+    Buffer,
+    LinearAverageCostPricing,
+    MapScore,
+    NegotiatingPlugin,
+    PropertyValueLerpScore,
+    ProposalScoringMixin,
+    RandomScore,
+    RejectIfCostsExceeds,
+    ScoringBuffer,
+)
 
-__all__ = ("DefaultProposalManager",)
+__all__ = (
+    "DefaultProposalManager",
+    "BlacklistProviderIdPlugin",
+    "Buffer",
+    "AddChosenPaymentPlatform",
+    "BlacklistProviderIdNegotiator",
+    "NegotiatingPlugin",
+    "RejectIfCostsExceeds",
+    "MapScore",
+    "ProposalScoringMixin",
+    "LinearAverageCostPricing",
+    "PropertyValueLerpScore",
+    "RandomScore",
+    "ScoringBuffer",
+)
