+++ golem/managers/proposal/plugins/scoring/__init__.py
@@ -1,0 +1,15 @@
+from golem.managers.proposal.plugins.scoring.map import MapScore
+from golem.managers.proposal.plugins.scoring.mixins import ProposalScoringMixin
+from golem.managers.proposal.plugins.scoring.pricings import LinearAverageCostPricing
+from golem.managers.proposal.plugins.scoring.property_value_lerp import PropertyValueLerpScore
+from golem.managers.proposal.plugins.scoring.random import RandomScore
+from golem.managers.proposal.plugins.scoring.scoring_buffer import ScoringBuffer
+
+__all__ = (
+    "MapScore",
+    "ProposalScoringMixin",
+    "LinearAverageCostPricing",
+    "PropertyValueLerpScore",
+    "RandomScore",
+    "ScoringBuffer",
+)
