+++ golem/managers/proposal/plugins/scoring/mixins.py
